requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
azure-core==1.29.6
azure-storage-blob==12.19.0
Babel==2.12.1
cachetools==4.2.4
certifi==2023.7.22
cffi==1.16.0
channels==3.0.4
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.2.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cryptography==41.0.5
cycler==0.12.1
daphne==3.0.2
distro==1.8.0
Django==4.0.3
django-autoslug==1.9.9
django-cors-headers==3.11.0
django-countries==7.5.1
django-crispy-forms==2.0
django-rest-swagger==2.2.0
django-shortuuidfield==0.1.3
django-storages==1.14.2
django-widget-tweaks==1.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.7
fonttools==4.47.2
gevent==23.9.1
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
MouseInfo==0.1.3
msgpack==1.0.7
numpy==1.26.3
oauthlib==3.2.2
openai==1.3.8
openapi-codec==1.3.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.1
Pillow==10.0.0
progressbar2==4.2.0
protobuf==4.23.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pynput==1.7.6
pyOpenSSL==23.3.0
pyparsing==3.1.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-decouple==3.8
python-utils==3.7.0
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.0
rsa==4.9
service-identity==23.1.0
setuptools==68.2.2
shortuuid==1.0.11
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==6.7.0
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
uritemplate==3.0.1
urllib3==2.0.7
websocket==0.2.1
websocket-client==1.7.0
websockets==12.0
wget==3.2
wheel==0.41.3
zope.event
zope.interface
